+++ src/EtherSharp/ABI/Types/AbiTypes.Bytes.cs
@@ -2,19 +2,20 @@
 using System.Buffers.Binary;
 
 namespace EtherSharp.ABI.Types;
+
 public static partial class AbiTypes
 {
     public class Bytes : DynamicType<byte[]>, IPackedEncodeType
     {
-        public override uint PayloadSize => (((uint) Value.Length + 31) / 32 * 32) + 32;
+        public override int PayloadSize => ((Value.Length + 31) / 32 * 32) + 32;
         /// <inheritdoc/>
         public int PackedSize => Value.Length;
 
         internal Bytes(byte[] value) : base(value) { }
 
-        public override void Encode(Span<byte> metadata, Span<byte> payload, uint payloadOffset)
+        public override void Encode(Span<byte> metadata, Span<byte> payload, int payloadOffset)
         {
-            BinaryPrimitives.WriteUInt32BigEndian(metadata[28..32], payloadOffset);
+            BinaryPrimitives.WriteUInt32BigEndian(metadata[28..32], (uint) payloadOffset);
             BinaryPrimitives.WriteUInt32BigEndian(payload[28..32], (uint) Value.Length);
 
             Value.CopyTo(payload[32..]);
@@ -22,7 +23,7 @@ public static partial class AbiTypes
         void IPackedEncodeType.EncodePacked(Span<byte> buffer)
             => Value.CopyTo(buffer);
 
-        public static ReadOnlySpan<byte> Decode(ReadOnlySpan<byte> bytes, uint metaDataOffset)
+        public static ReadOnlySpan<byte> Decode(ReadOnlySpan<byte> bytes, int metaDataOffset)
         {
             uint bytesOffset = BinaryPrimitives.ReadUInt32BigEndian(bytes[((int) metaDataOffset + 28)..((int) metaDataOffset + 32)]);
